FST-80 - refactor naming convention, update menuitems
