+++ src/App.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import {createTheme, ThemeProvider} from "@mui/material";
-import Main from "./main/main";
+import Main from "./main/Main";
 import {BrowserRouter} from "react-router-dom";
 
 const theme = createTheme({
