+++ src/interfaces/Route.ts
@@ -4,5 +4,4 @@ export default interface IRoute {
     path: string;
     component: any;
     props?: any;
-    icon?: string;
 }
